Testing: Move main_window fixture to its corresponding conftest file
This makes easier to contribute to that fixture separately from the
tests that use it.
